Update to CasperCheck's README
Adding an update to CasperCheck's README, to clarify that with Casper
9.8 and later on Macs running 10.7 and later, CasperCheck will be
checking for the presence of the "/usr/local/bin/jamf" symlink.

If the symlink is missing, CasperCheck's interpretation is that the
symlink's absence means the Casper agent is not working properly and
needs to be reinstalled using the QuickAdd installer stored on the Mac.